+++ src/app/progress/Milestones.tsx
@@ -3,7 +3,8 @@ import { DimStore } from 'app/inventory/store-types';
 import {
   DestinyProfileResponse,
   DestinyMilestone,
-  DestinySeasonDefinition
+  DestinySeasonDefinition,
+  DestinySeasonPassDefinition
 } from 'bungie-api-ts/destiny2';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import WellRestedPerkIcon from './WellRestedPerkIcon';
@@ -32,6 +33,7 @@ export default function Milestones({
   const profileMilestones = milestonesForProfile(defs, profileInfo, store.id);
   const characterProgressions = profileInfo?.characterProgressions?.data?.[store.id];
   const season = currentSeason(defs);
+  const seasonPass = currentSeasonPass(defs, season);
 
   const milestoneItems = [
     ...milestonesForCharacter(defs, profileInfo, store),
@@ -46,6 +48,7 @@ export default function Milestones({
           defs={defs}
           characterProgressions={characterProgressions}
           season={season}
+          seasonPass={seasonPass}
           profileInfo={profileInfo}
         />
       )}
@@ -127,5 +130,14 @@ function currentSeason(defs: D2ManifestDefinitions): DestinySeasonDefinition | u
       season.endDate &&
       new Date(season.startDate).getTime() < Date.now() &&
       new Date(season.endDate).getTime() > Date.now()
+  );
+}
+
+function currentSeasonPass(
+  defs: D2ManifestDefinitions,
+  season: DestinySeasonDefinition | undefined
+): DestinySeasonPassDefinition | undefined {
+  return Object.values(defs.SeasonPass.getAll()).find(
+    (seasonPass) => seasonPass.rewardProgressionHash === season?.seasonPassProgressionHash
   );
 }